requirements.txt
streamlit
requests
google-cloud-bigquery
plotly
pandas
numpy
watchdog
db-dtypes
pandas-gbq
